Fixed bugs in the re-written split system.
